+++ app/extensions.py
@@ -3,9 +3,12 @@ from flask_mail import Mail
 from flask_migrate import Migrate
 from flask_wtf.csrf import CSRFProtect
 from flask_login import LoginManager
+from flask_limiter import Limiter
+from flask_limiter.util import get_remote_address
 
 db = SQLAlchemy()
 mail = Mail()
 migrate = Migrate()
 csrf = CSRFProtect()
 login_manager = LoginManager()
+limiter = Limiter(key_func=get_remote_address)